hotel_backend-master/src/main/java/tech/selmefy/hotel/service/hotelservice/HotelServiceService.java
package tech.selmefy.hotel.service.hotelservice;

import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;
import tech.selmefy.hotel.controller.hotelservice.dto.HotelServiceDTO;
import tech.selmefy.hotel.mapper.HotelServiceMapper;
import tech.selmefy.hotel.repository.hotelservice.HotelServiceRepository;

import java.util.List;

@Service
@AllArgsConstructor

public class HotelServiceService {

    public final HotelServiceRepository hotelServiceRepository;

    public List<HotelServiceDTO> getAllHotelServices() {
        return hotelServiceRepository.findAll()
                .stream()
                .map(HotelServiceMapper.INSTANCE::toDTO)
                .toList();
    }

    public HotelServiceDTO getHotelServiceById(short id) {
        return hotelServiceRepository.findById(id).map(HotelServiceMapper.INSTANCE::toDTO).orElseThrow();
    }
}
